WL set effects
